Refactor API routes in web.php to enhance structure and add LDAP functionality; include new endpoints for investor authentication and asset outstanding details; add composer zip file for vendor updates.

+++ CPM_Service/app/Services/Handlers/Finance/AssetLiabilityService.php
@@ -23,8 +23,8 @@ class AssetLiabilityService
         $asset = $this->assetRepo->totalAsset($investorId, $endDate);
         $liability = $this->liabilityRepo->totalLiability($investorId, $endDate);
 
-        $totalAsset = $asset ?? 0;
-        $totalLiability = $liability ?? 0;
+        $totalAsset = (float) ($asset ?? 0);
+        $totalLiability = (float) ($liability ?? 0);
 
         return [
             'asset' => $totalAsset,
